[pce] add a special IO register (region?) to allow logfile writing from program.
the first zero-terminated string sent to $1c00 I/O register is the filename. This can be a relative filename in which case it is relative from the current directory you ran from (this does not mean the directory the mednafen binary is located). All subsequent writes to the register will be directly dumped as binary/ASCII to the logfile specified.