[debugger] automatically add a BRK opcode-breakpoint at startup
Ideally this would ONLY affect PCE -- the system I'm targeting, but since I do not know how to do that, it just globally adds the breakpoint. FYI, "0" is the opcode for BRK on Huc6280. The patch just adds the opcode to the debugger's OpBreakpoints std::string and then calls "UpdateBreakpoints" with it. I learned to do that from debugging. I'm not sure if I'm missing extra steps but everything seems to be working fine. On a brief initial test the Mednafen debugger's branch tracing appears accurate after a BRK break.

The whole reason why I'm adding this feature is because I am writing a game which sometimes incurs bugs that cause BRK to execute.. It's a very common occurrence in a system crash on the PCE in my experience.. And right now, I have a bug that I have no idea how to re-trigger.. So I'm hoping that I'll accidently stumble upon it again with my new auto-BRK-breakpoint assigned, and be able to get some idea of what led to the problem (as long as the branch trace can contain the issue).